Changes per PR feedback
