--- app/src/main/java/com/amaze/filemanager/database/typeconverters/JsonTypeConverter.kt
@@ -1,47 +1,0 @@
-/*
- * Copyright (C) 2014-2022 Arpit Khurana <arpitkh96@gmail.com>, Vishal Nehra <vishalmeham2@gmail.com>,
- * Emmanuel Messulam<emmanuelbendavid@gmail.com>, Raymond Lai <airwave209gt at gmail.com> and Contributors.
- *
- * This file is part of Amaze File Manager.
- *
- * Amaze File Manager is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package com.amaze.filemanager.database.typeconverters
-
-import androidx.room.TypeConverter
-import org.json.JSONObject
-
-/**
- * [TypeConverter] between [JSONObject] as object and String representation.
- */
-object JsonTypeConverter {
-    /**
-     * Convert from [JSONObject] to string.
-     */
-    @JvmStatic
-    @TypeConverter
-    fun fromJsonObject(value: JSONObject): String {
-        return value.toString()
-    }
-
-    /**
-     * Convert from string to [JSONObject].
-     */
-    @JvmStatic
-    @TypeConverter
-    fun fromJsonString(value: String): JSONObject {
-        return JSONObject(value)
-    }
-}
