+++ ftpserver/src/main/java/com/amaze/filemanager/ftpserver/service/FtpServerEvent.kt
@@ -1,0 +1,14 @@
+package com.amaze.filemanager.ftpserver.service
+
+/**
+ * Events broadcast when FTP server state changes.
+ */
+sealed class FtpServerEvent {
+    data object Started : FtpServerEvent()
+
+    data object StartedFromTile : FtpServerEvent()
+
+    data object Stopped : FtpServerEvent()
+
+    data object FailedToStart : FtpServerEvent()
+}
